Relocate into Features layer and split classes into own files
